move generator to src
